fix: correct C++20 std::format incompatibilities
- callback_list_impl.h: fix malformed LogErr call (semicolon inside
  parens, try_clear() stuck in argument list)
- file_cache.cpp: add .string() to all std::filesystem::path log args
  (std::filesystem::path is only formattable in C++23, not C++20)
- fs_utils.cpp: restore <sstream> include (removed by the iomanip
  cleanup but still needed for std::stringstream)
- system_impl.cpp: cast MAV_TYPE enum to int before formatting
  (C enums have no std::formatter specialization in C++20)

+++ cpp/src/mavsdk/core/system_impl.cpp
@@ -973,7 +973,7 @@ SystemImpl::make_command_ardupilot_mode(FlightMode flight_mode, uint8_t componen
         }
 
         default:
-            LogErr("Cannot translate flight mode to ArduPilot mode, for MAV_TYPE: {}", _vehicle_type);
+            LogErr("Cannot translate flight mode to ArduPilot mode, for MAV_TYPE: {}", static_cast<int>(_vehicle_type));
             MavlinkCommandSender::CommandLong empty_command{};
             return std::make_pair<>(MavlinkCommandSender::Result::UnknownError, empty_command);
     }
